refactor: split API modules into explicit bounded contexts
- Reorganize from feature-based to domain-based structure
- Create domains: identity, journeys, discovery, resonance,
  messaging, events, taste-signals, moderation, analytics, payments
- Preserve current startup behavior and route registration
- Update all import paths

Closes Issue #7
