+++ apps/api/src/domains/analytics/index.ts
@@ -1,0 +1,8 @@
+// Analytics domain - Usage tracking, insights
+// TODO: Implement in future sprint
+
+import { Router } from 'express';
+
+const analyticsRouter = Router();
+
+export default analyticsRouter;
