+++ apps/api/src/domains/discovery/discovery.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
-import { IDjProfileDocument } from '../profiles/dj.model';
-import DjProfile from '../profiles/dj.model';
-import User from '../users/user.model';
+import { IDjProfileDocument } from './dj.model';
+import DjProfile from './dj.model';
+import User from '../identity/user.model';
 
 const serializeDj = (
   profile: Pick<
