+++ apps/api/src/domains/discovery/discovery.routes.ts
@@ -1,9 +1,4 @@
 import { Router } from 'express';
-import { listDjs } from './discovery.controller';
-
-const discoveryRouter = Router();
-
-discoveryRouter.get('/djs', listDjs);
 import { getDjProfile } from './discovery.controller';
 
 const discoveryRouter = Router();
