+++ apps/api/src/domains/discovery/discovery.validation.ts
