+++ apps/api/src/domains/discovery/index.ts
@@ -1,0 +1,1 @@
+export { default as discoveryRouter } from './discovery.routes';
