+++ apps/api/src/domains/events/index.ts
@@ -1,0 +1,8 @@
+// Events domain - Event management and attendance
+// TODO: Implement in future sprint
+
+import { Router } from 'express';
+
+const eventsRouter = Router();
+
+export default eventsRouter;
