+++ apps/api/src/domains/identity/index.ts
@@ -1,0 +1,1 @@
+export { default as identityRouter } from './auth.routes';
