+++ apps/api/src/domains/journeys/index.ts
@@ -1,0 +1,1 @@
+export { default as journeysRouter } from './bookings.routes';
