+++ apps/api/src/domains/messaging/index.ts
@@ -1,0 +1,8 @@
+// Messaging domain - Communication between users
+// TODO: Implement in future sprint
+
+import { Router } from 'express';
+
+const messagingRouter = Router();
+
+export default messagingRouter;
