+++ apps/api/src/domains/moderation/index.ts
@@ -1,0 +1,8 @@
+// Moderation domain - Content moderation, reporting
+// TODO: Implement in future sprint
+
+import { Router } from 'express';
+
+const moderationRouter = Router();
+
+export default moderationRouter;
