+++ apps/api/src/domains/payments/index.ts
@@ -1,0 +1,1 @@
+export { default as paymentsRouter } from './payments.routes';
