+++ apps/api/src/domains/payments/stellar-gateway.ts
