+++ apps/api/src/domains/resonance/index.ts
@@ -1,0 +1,8 @@
+// Resonance domain - Taste signals, reactions, matching algorithms
+// TODO: Implement in future sprint
+
+import { Router } from 'express';
+
+const resonanceRouter = Router();
+
+export default resonanceRouter;
