+++ apps/api/src/domains/taste-signals/index.ts
@@ -1,0 +1,8 @@
+// Taste Signals domain - Music preferences, listening history
+// TODO: Implement in future sprint
+
+import { Router } from 'express';
+
+const tasteSignalsRouter = Router();
+
+export default tasteSignalsRouter;
