+++ apps/api/src/index.ts
@@ -1,23 +1,23 @@
+import express from 'express';
+import cors from 'cors';
 import connectDB from './config/db';
 import { apiEnv } from './config/env';
-import authRouter from './modules/auth/auth.routes';
-import paymentsRouter from './modules/payments/payments.routes';
-import bookingsRouter from './modules/bookings/bookings.routes';
-import apiV1Router from './routes/api-v1.router';
+import { identityRouter } from './domains/identity';
+import { journeysRouter } from './domains/journeys';
+import { discoveryRouter } from './domains/discovery';
+import { paymentsRouter } from './domains/payments';
 import { notFoundHandler } from './middleware/not-found.middleware';
 import { errorHandler } from './middleware/error.middleware';
-import discoveryRouter from './modules/discovery/discovery.routes';
-import profilesRouter from './modules/profiles/profiles.routes';
 
 const app = express();
 const port = apiEnv.port;
 
 app.use(cors({ origin: apiEnv.corsOrigin }));
 app.use(express.json());
-app.use('/auth', authRouter);
-app.use('/payments', paymentsRouter);
+app.use('/auth', identityRouter);
+app.use('/bookings', journeysRouter);
 app.use('/discover', discoveryRouter);
-app.use('/profiles', profilesRouter);
+app.use('/payments', paymentsRouter);
 
 connectDB();
